+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* [Small fix for context menu on mobile and item leave event](https://github.com/flower-platform/react-timeline-10000/pull/86/files)
 * [Update comment for rct9k-items-inner style class](https://github.com/flower-platform/react-timeline-10000/pull/88/files)
 
 ## v2.7.5